frontend/services/jwt_service/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.1.1
PyJWT==2.9.0
redis==5.2.0
cryptography==43.0.3
locust==2.17.0

